requirements.txt
discord.py>=2.3.2
aiohttp>=3.9.0
gitpython>=3.1.40
click>=8.1.7
python-dotenv>=1.0.0
